initramfs-crypt-hook: Factor out tpm2_functions script
This is supposed to handle differences of tpm2-tools version 3 as found
in buster. We start with the tpm discovery. Later on, persistence of the
initial passphrase will be added.

Note that we intentionally do not stop the search after the first hit in
order to remain backward-compatible with older versions of the hook.
This means we will continue to pick the last working tpm device in case
there are more than one.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>