create apt directory as needed by apt-daily
The apt daily service takes care of daily checking for updates. This
currently fails to start because the /var/lib/apt directory is not
available (it is created by running a manual "apt update", but if not
done explicitly, the service will fail).

While this service is not needed on a ro-rootfs system, it is provided
by the apt package which is installed and automatically enabled.
Disabling the service is an independent topic.

By not providing this directory, the system comes up in a "degraded"
state after the first time the service is executed. This is problematic,
as other logic (e.g. the blessing after an update) might also depend on
this state.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>