immutable-rootfs: disable daily package update checks
When running on an immutable rootfs, we don't allow package updates.
Nonetheless, apt tries to update the daily package list and upgrade some
packages. This is triggered by systemd timers, which are both manually
registered in /etc, as well as deployed via systemd vendor presets.
While it anyways makes no sense to run the update, it also might fail
for various reasons, rendering the system "degraded" after some time.

We disable the timers and also disable the dpkg database backup, as the
database itself is also immutable.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: bumped recipe version]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>